+++ lib/kinesis_client.rb
@@ -1,7 +1,7 @@
-require 'securerandom'
-require 'aws-sdk-kinesis'
-require_relative 'nypl_avro'
-require_relative 'errors'
+require "securerandom"
+require "aws-sdk-kinesis"
+require_relative "nypl_avro"
+require_relative "errors"
 # Model representing the result message posted to Kinesis stream about everything that has gone on here -- good, bad, or otherwise.
 
 class KinesisClient
@@ -16,13 +16,15 @@ class KinesisClient
     @automatically_push = @config[:automatically_push] == false ? false : true
     @client_options = set_config(config)
     @client = Aws::Kinesis::Client.new @client_options
+    @batch_count = 0
+    @records = []
+    @automatically_push = !(@config[:automatically_push] == false)
+    @client_options = config[:profile] ? { profile: config[:profile] } : {}
+    @client = Aws::Kinesis::Client.new(@client_options)
 
-    if config[:schema_string]
-      @avro = NYPLAvro.by_name(config[:schema_string])
-    end
+    @avro = NYPLAvro.by_name(config[:schema_string]) if config[:schema_string]
 
-    @shovel_method = @batch_size > 1 ? :push_to_batch : :push_record
-
+    @shovel_method = @batch_size > 1 ? :push_to_records : :push_record
   end
 
   def set_config(config)
@@ -47,18 +49,18 @@ class KinesisClient
       data: message,
       partition_key: partition_key
     }
-
   end
 
   def <<(json_message)
     send(@shovel_method, json_message)
   end
 
+#This method is broken
   def push_record(json_message)
-    record = convert_to_record json_message
+    record = convert_to_record(json_message)
     record[:stream_name] = @stream_name
 
-    @client.put_record record
+    @client.put_record(record)
 
     return_hash = {}
 
@@ -67,21 +69,19 @@ class KinesisClient
       return_hash["message"] = json_message, resp
       $logger.info("Message sent to #{config[:stream_name]} #{json_message}, #{resp}") if $logger
     else
-      $logger.error("message" => "FAILED to send message to HoldRequestResult #{json_message}, #{resp}.") if $logger
-      raise NYPLError.new json_message, resp
+      $logger.error("message" => "FAILED to send message to #{@stream_name} #{json_message}, #{resp}.") if $logger
+      raise(NYPLError.new(json_message, resp))
     end
     return_hash
   end
 
-  def push_to_batch(json_message)
+  def push_to_records(json_message)
     begin
-      @batch << convert_to_record(json_message)
+      @records << convert_to_record(json_message)
     rescue AvroError => e
       $logger.error("message" => "Avro encoding error #{e.message} for #{json_message}")
     end
-    if @automatically_push && @batch.length >= @batch_size
-      push_records
-    end
+    push_records if @automatically_push && @records.length >= @batch_size
   end
 
   def push_batch(batch)
@@ -90,23 +90,31 @@ class KinesisClient
       stream_name: @stream_name
     })
 
-    $logger.debug("Received #{resp} from #{@stream_name}")
-
-    return_message = {
-      failures: resp.failed_record_count,
-      error_messages: resp.records.map {|record| record.error_message }.compact
-    }
-
-    $logger.info("Message sent to #{config[:stream_name]} #{return_message}") if $logger
-
-    return {
-      "code": "200",
-      "message": return_message.to_json
-    }
+    if resp.failed_record_count > 0
+      failure_message = {
+        failures: resp.failed_record_count,
+        failures_data: filter_failures(resp)
+      }
+      $logger.warn("Batch sent to #{config[:stream_name]} with failures: #{failure_message}")
+    else
+      $logger.info("Batch sent to #{config[:stream_name]} successfully")
+    end
   end
 
   def push_records
-    @batch.each_slice(@batch_size) {|slice| push_batch slice}
-    @batch = []
+    if @records.length > 0
+      @records.each_slice(@batch_size) do |slice|
+        push_batch(slice)
+        @batch_count += 1
+      end
+      @records = []
+      @batch_count = 0
+    end
+  end
+
+  def filter_failures(resp)
+    resp.records.filter_map.with_index do |record, i|
+      avro.decode(@records[i + @batch_size * @batch_count]) if record.responds_to?(:error_message)
+    end
   end
 end