+++ lib/kms_client.rb
@@ -1,5 +1,5 @@
-require 'aws-sdk-kms'
-require 'base64'
+require "aws-sdk-kms"
+require "base64"
 
 class KmsClient
   @@kms = nil
@@ -11,15 +11,15 @@ class KmsClient
 
   def decrypt(cipher)
     # Assume value is base64 encoded:
-    decoded = Base64.decode64 cipher
-    decrypted = @kms.decrypt ciphertext_blob: decoded
+    decoded = Base64.decode64(cipher)
+    decrypted = @kms.decrypt(ciphertext_blob: decoded)
     decrypted[:plaintext]
   end
 
   def self.aws_kms_client(options)
     params = {
-      region: 'us-east-1',
-      stub_responses: ENV['APP_ENV'] == 'test'
+      region: "us-east-1",
+      stub_responses: ENV["APP_ENV"] == "test"
     }.merge(options)
     @@kms = Aws::KMS::Client.new(params) if @@kms.nil?
     @@kms